Initial VQVAE including code from prac2 VAE
The code contains code from prac2 VAE project but the current progress for the VQVAE is in the load_data function which contains the data splitting into train and validation sets. Additionally, changing the dimension of the data from tf.Dataset object to a flat images array.
other code is placeholder from prac2 for now.