Added initial models for Encoder, Decoder, VQVAE
Code clean-up: Removed unnecessary code from VAE of prac2.
Added Encoder model (this involves the initial network and the model summary.
Added Decoder model (this involves the network and the model summary)
Added VQVAE model (this involves both Encoder and Decoder layers with the custom layer of the vector quantizer which is to be implemented at a later stage.
Added simple print statement for printing the original and reconstruction of the images.