Refine Model Parameters, Begin Dataset Loading
Reduce some of the bounding box and other parameters by a factor of 10 to reflect our reduced number of classes versus the COCO dataset. Begin basic loading of our data into a training and validation set, with a test set to be done later (with care taken not to contaminate). Also doing a simple image flip augmentation. Next to do is create custom class with a load_image() and load_mask() method as required for Mask R-CNN.