Change from Mask R-CNN to YOLOv5
Remove all Mask R-CNN code and begin creation of simple YOLO class using PyTorch with the loaded medium sized model trained on COCO. Rename data loader class to setup class, that is now its own program with main. This program should be run to move the training data into subdirectories and randomly shuffle it. Additionally, we must create a YAML file and text files for each image to be able to pass our bounding boxes into the model, as YOLO does not use masks. Update .gitignore to not include Mac specific file and all of the training data in the subdirectories. Run the first inference, which bounded quite well but of course for the wrong class (cat).