Added VectorQuantizer and training history plot
Added VectorQuantizer model implementation essentially the calculations are made in the call function however, the avg_probs and perplexity are yet to be integrated.
Some code clean-up and debugging is added to figure out issues with a layer producing nan. Currently working on the gradient training.
Added plot for the training loss using the history tracking.