Added the trainer and testing output
Added the trainer but it is currently not working well as the images produced are not clear and some layer is have issues with the gradient as the gradient is nan at that point. I have split the gradient as the tutor advised and I added some sample testing code to check the outputs.
The vectorQuantizer layer is the same from last commit but moved up.