Compress JPEGs to 1024
Resize all of the images with the maximum axis being 1024. This has brought the dataset size from 3.3GB to 170MB. Also ensure we are closing our files correctly by using the with operator, as file IO objects sit in unmanaged memory and will not be collected by the RC garbage collector.