Added a sample residual stack and different encoder decoders
Added a sample residual stack currently it does not function as it hasn't been yet integrated to the encoder and decoder layers.
The encoder and decoder layers has been changes as recommend by the tutor but the images still not clear and as such they are still being developed.
The commented out code consists of testing VectorQuantizer class and loss methods. Those items are not used in the code and are just for testing until I get a working version.