Change Image Dimensions To 640, Augment Data
Change the resize default dimensions to 640 on the longest axis, which is the default for YOLOv5. Rewrite some of the isics_data_setup file to not copy the mask data into the new directories, as we don't need it, and also to generate augmented data. For each file we do a 90, 180, and 270 degrees rotation, as well as a horizontal and vertical flip from the original image.