Added the driver and model trainer
This new file contains the driver of the models and trainer model used in a previous commit. Changes include adding adding each layer separately and adding a pre-vq_vae layer. The model is compiled in sequence just to view the summary but it is not used. The custom train_step would train all the layers on it's own and calculate the loss on each step. Finally, losses are printed and model training history is saved. Furthermore, I added training hyperparameters that help adjust the model layers and propagate change easier compared to pervious implementation where it was just numbers and not variables.