Added the missing plotting functions
Added the missing plotting for the predicted data images and the history. I had to expand the dimensions of the data as the network input contains the first number of samples dim but then I take and output and squeeze it to remove the extra dimension when plotting. 3 dims -> 4 dims - > 3 dims for the testing data.