Fixed big loss issue and plots
I finally fixed the issue of the losses being very large (thanks to Siyu) by removing the division by the data variance. The data variance appeared to be very small and causing the loss to maximize.
I have added 10 plot samples instead of just 1 in order to add them for the readme.