Created Improved Unet model.
The model.ipynb file now contains what should be an Improved UNet model
based on the paper stipulated in the task sheet. While it still contains
the original UNet I created as a test and to get experience with UNet's
that will be removed once the Improved UNet is shwown to be working.

The current format contains all functions needed to run the function
that creates the UNet.

The next step is again try to get the keras plot_model() working on
windows.