Added comments and cleaned up code in driver.py
Added comments to both data_reader and driver files.
In driver file I made changes to the data_variance variable by setting it to 0 as it is not used for the current model due to the data large variance.
Trained the data for 100 epochs and removed the use of numpy library by replacing numpy.zeros with tf.zeros.