Fixed an issue with caused by removing numpy.zeros
Remvoing numpy.zeors and replacing it with tf.zeros caused the issue of returning an Eager tensor object instead of an array and as such was throwing not assignable error. To fix I converted the tf.zeros object into an array after it was created in both driver.py for the priors and in model.py for the PixelCNN model.