Updated test visualisation and implemented test set testing.
The functions for visualising the predicted masks compared to real masks
were updated and integrated into the main function. The visualisation
function now accurately shows two test images, their predicted masks and
their ground truth masks.

A function that avaluates the model on the entire test dataset was
implemented and integrated into the main function. This function
performs an evaluation on the entire 20% of all images that were
reserved for testing. It achieved a 0.89 Dice Coefficient score.

Due to these results, further refinement of the model, augmentation, and
training process is not needed.

Todo: Improve annotations on test visualisation output and look at why
the data the model['accuracy'] is tracking seems to be not relevant.