Update comment of train.py.
you can jump the train model, if only want to use the model.

Actually the model is trained based on a pre-trained 15 epoch one. So please set a higher number of epoch if you want to train by your own. If you want to run to prove my work is sufficient. 

The train result from this part will be shown in the README part.