Edited VQVAE Model to fix bugs to allow model fitting and validation error to be calculated.
dataset.py updated so channels is consistent with model

Model training code was included (Trained on Collab), as well as plotting code for the error, plotting code for example reconstructions, and finding the mean SSIM for the test data. Currently mean SSIM is 90% on test dataset. Utils code not in functions yet