increased UNET num trainable param and fixes
I have added in another time to save the model (but to a separate file for a back up). I let it train overnight and it show how crashed, when I tried loading the model it was corrupted. This is because it would've crashed whilst saving the model. I have created a separate instance of it saving the model. But at a different time to the first save. That way if one becomes corrupted I can load up from a 5 epochs ago. Also have fixed my de noising for my predict file. Which can show images of stuff that look similar to brains.