Immediately Preprocess Data
Changed download_dataset so that it automatically processes the data. This then meant class variable names were changed, which had a flow-on effect through the class and separated the processing of the image and the truth as I would otherwise have to change my dataset into a tuple for the data and truth sets.