Tensor Slices Zero Error Fixed
When running, there was a problem with using from_tensor_slices twice as it had one of the dimensions as none, and we couldn’t combine two tensors like that. It was changed so when the tensor slices are first called it forms two-dimensional tensors. This means the pre_processing had to be combined as passing it into the map with include both the image and truth